Making it before() middleware
